+++ components/channels-dialog/confirm.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable max-len */
 /**
  * confirm 是将 Dialog 函数式编程的结果。
  * 1. confirm 是为了让使用者方便地函数式调用组件；因此 confirm 方法的 visible 状态是不交给外部处理的；
@@ -8,9 +7,18 @@ import * as React from "react"
 import * as ReactDOM from "react-dom"
 import classNames from "classnames"
 import Dialog, { IDialogProps } from "./Dialog"
+import Icon from "../icon"
 import "./style"
 
 const prefix = "adui-channels-dialog"
+
+const ICON_INTENTS = {
+  danger: "alert-circle" as "alert-circle",
+  info: "info-circle" as "info-circle",
+  success: "tick-circle" as "tick-circle",
+  warning: "warning" as "warning",
+}
+
 export interface IConfirmProps extends IDialogProps {
   content?: React.ReactNode
   contentTitle?: React.ReactNode
@@ -63,37 +71,11 @@ const Confirm = (props: IConfirmProps) => {
             }
           )}
         >
-          {intent === "danger" ? (
-            <svg
-              width="48"
-              height="48"
-              viewBox="0 0 48 48"
-              fill="none"
-              className={`${prefix}-intentIcon`}
-            >
-              <path
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M4 24C4 35.0457 12.9543 44 24 44C35.0457 44 44 35.0457 44 24C44 12.9543 35.0457 4 24 4C12.9543 4 4 12.9543 4 24ZM41.6 24C41.6 33.7202 33.7202 41.6 24 41.6C14.2798 41.6 6.4 33.7202 6.4 24C6.4 14.2798 14.2798 6.4 24 6.4C33.7202 6.4 41.6 14.2798 41.6 24ZM25.3184 12.8623L25.1426 26.9834H22.8574L22.6816 12.8623H25.3184ZM22.3301 32.4766C22.3301 33.4141 23.0625 34.1465 24 34.1465C24.9521 34.1465 25.6699 33.4141 25.6699 32.4766C25.6699 31.5244 24.9521 30.8066 24 30.8066C23.0625 30.8066 22.3301 31.5244 22.3301 32.4766Z"
-                fill="#fa5151"
-              />
-            </svg>
-          ) : (
-            <svg
-              width="48"
-              height="48"
-              viewBox="0 0 48 48"
-              fill="none"
-              className={`${prefix}-intentIcon`}
-            >
-              <path
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M4 24C4 35.0457 12.9543 44 24 44C35.0457 44 44 35.0457 44 24C44 12.9543 35.0457 4 24 4C12.9543 4 4 12.9543 4 24ZM41.6 24C41.6 33.7202 33.7202 41.6 24 41.6C14.2798 41.6 6.4 33.7202 6.4 24C6.4 14.2798 14.2798 6.4 24 6.4C33.7202 6.4 41.6 14.2798 41.6 24ZM25.2 20V34H22.8V20H25.2ZM24 18C25.1046 18 26 17.1046 26 16C26 14.8954 25.1046 14 24 14C22.8954 14 22 14.8954 22 16C22 17.1046 22.8954 18 24 18Z"
-                fill={intent === "info" ? "#10aeff" : "#000"}
-              />
-            </svg>
-          )}
+          <Icon
+            className={`${prefix}-intentIcon`}
+            size={48}
+            icon={ICON_INTENTS[intent]}
+          />
           <div className={`${prefix}-intentContent`}>
             {!!contentTitle && (
               <div className={`${prefix}-intentTitle`}>{contentTitle}</div>
@@ -135,15 +117,16 @@ const confirm = (config: IConfirmProps) => {
    * 关闭时一定要执行 afterClose，Dialog 自身会去 destroy，但是那个空的 div 需要在这里 destory。
    */
   function onClose(cb: () => void) {
+    let cbReturn: any
+    if (cb) {
+      cbReturn = cb()
+    }
     currentConfig = {
       ...currentConfig,
       afterClose,
-      visible: false,
+      visible: cbReturn === false,
     }
     render(currentConfig)
-    if (cb) {
-      cb()
-    }
   }
 
   currentConfig = {